requirements.txt
Flask==2.3.3
gunicorn==21.2.0
google-generativeai==0.3.1
python-dotenv==1.0.0
Flask-Session==0.5.0
Werkzeug==2.3.7
requests==2.31.0